src/transport/tests/requirements-test.txt
pytest==8.0.0
pytest-asyncio==0.23.5
aiohttp==3.9.3
websockets==12.0
httpx==0.26.0
PyJWT==2.8.0
python-dotenv==1.0.0
fastapi==0.110.0
uvicorn==0.27.1 